[FEATURE] Add new possibilities to modify DefaultUploadFolder settings
Before this change the resolving of the default upload folder
was done in BackendUserAuthentication. As new users wished for
additional ways to configure the default upload folder (e.g.
per page, per record, etc.), it started to not make sense that
the logic is present in the backend user.

The DefaultUploadFolderResolver uses the same logic as before
including a new `options.defaultUploadFolder` PageTSconfig setting,
allowing to change this setting on a per-page level.

In addition, a new PSR-14 event
`AfterDefaultUploadFolderWasResolvedEvent` is introduced
receiving the same arguments as the hook in BackendUserAuthentication named

$GLOBALS['TYPO3_CONF_VARS']['SC_OPTIONS']['t3lib/class.t3lib_userauthgroup.php']['getDefaultUploadFolder']

which serves the same purpose. This way, the upload folder
can also be modified by PHP developers.

Resolves: #83608
Releases: main
Change-Id: Ida5c7c559606bbd0e607debdc6b0a781a4c69cc2
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/76550
Tested-by: core-ci <typo3@b13.com>
Tested-by: Benni Mack <benni@typo3.org>
Tested-by: Georg Ringer <georg.ringer@gmail.com>
Reviewed-by: Benni Mack <benni@typo3.org>
Reviewed-by: Georg Ringer <georg.ringer@gmail.com>