[TASK] Avoid additional nesting complexity in typo3-backend-column-selector-button
The custom element typo3-backend-column-selector-button is a
valid HTML and should be used as such. There is no need to nest
buttons/links inside the element to add CSS styling.

We are adding keyboard events to react on the "Enter" and "Space"
keys to mimic the behavior of a button and setting defaults for the
role and tabindex to make it keyboard accessible. CSS classes are
now directly set to the element itself.

Resolves: #99080
Releases: main
Change-Id: I71dfa68174255a5d445af5fadbf1924b54a6b687
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/76593
Tested-by: core-ci <typo3@b13.com>
Tested-by: Andreas Fernandez <a.fernandez@scripting-base.de>
Reviewed-by: Andreas Fernandez <a.fernandez@scripting-base.de>
Tested-by: Benni Mack <benni@typo3.org>
Reviewed-by: Benni Mack <benni@typo3.org>