[TASK] Replace prophecy in EXT:core/Tests/Unit/Database/*
Note, that the combination of shouldBeCalled and shouldNotBeCalled
for the `quoteIdentifier` function in `QueryBuilderTest` has
been replaced with the `never` expectation only, since the tests
only test, if the star placeholder is not quoted.

Also 2 new mock classes have been introduced to cover the tests
in `AbstractRestrictionContainerTest`, where the mock object
must implement 2 interfaces, which is not mockable.

Resolves: #98792
Resolves: #98793
Resolves: #98794
Resolves: #98795
Resolves: #98796
Resolves: #98797
Resolves: #98798
Resolves: #98799
Resolves: #98800
Resolves: #98801
Resolves: #98802
Resolves: #98803
Resolves: #98804
Releases: main
Change-Id: I691c69f201e39cd8d6d580cbbb86fad4cb032c64
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/76666
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: core-ci <typo3@b13.com>
Reviewed-by: Stefan Bürk <stefan@buerk.tech>
Tested-by: Stefan Bürk <stefan@buerk.tech>