[BUGFIX] Do not mark the FAL-related Extbase models as internal anymore
The FAL-related Extbase models are documented in the "TYPO3 Explained"
documentation, and they are the recommended way to use FAL in an
Extbase context. So the following classes are not marked as `@internal`
anymore in order to not send mixed signals to extension developers:

- TYPO3\CMS\Extbase\Domain\Model\File
- TYPO3\CMS\Extbase\Domain\Model\FileReference
- TYPO3\CMS\Extbase\Domain\Model\Folder

There is no point in forcing extension authors to re-implement
those classes (and duplicating the code from the Core) when
the Core implementation is stable now and has proven to
work fine.

Some history for what has happened:

Those classes started out marked as experimental (when they were
indeed still experimental). Then, in a cleanup bulk change in
3888dae8f3f71c3f5e1db1c00bcbfe7d0d8fed2e5966e6533c9353b6d7485f32,
the classes were converted to use the `@internal` annotation,
even though at least the three FAL Extbase models should have
been made public API instead by that time.

This is where the usage of the FAL models is documented in
"TYPO3 Explained":
https://docs.typo3.org/m/typo3/reference-coreapi/main/en-us/ApiOverview/Fal/UsingFal/Frontend.html

Change-Id: I5e5dfcf292eabd57a080bed29572b7c1eecd9286
Resolves: #99126
Releases: main, 11.5
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/76674
Tested-by: core-ci <typo3@b13.com>
Tested-by: Benni Mack <benni@typo3.org>
Reviewed-by: Benni Mack <benni@typo3.org>