[BUGFIX] Make backdrop for EXT:recycler confirmations static
The confirmation in modal windows of EXT:recycler may run long
running tasks, depending on the database structure and amount of
data. It was possible to close the modal by clicking the backdrop,
while the actions still run which may lead to undefined behavior
when the recycler is used. The backdrop is now made static.

Resolves: #99181
Releases: main
Change-Id: I60cae362382bf249d4b6ac564f5134a2c782ef9e
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/76806
Tested-by: core-ci <typo3@b13.com>
Reviewed-by: Josef Glatz <josefglatz@gmail.com>
Reviewed-by: Andreas Fernandez <a.fernandez@scripting-base.de>
Reviewed-by: Benni Mack <benni@typo3.org>
Tested-by: Benni Mack <benni@typo3.org>
Tested-by: Andreas Fernandez <a.fernandez@scripting-base.de>
Tested-by: Josef Glatz <josefglatz@gmail.com>
Reviewed-by: Oliver Bartsch <bo@cedev.de>
Tested-by: Oliver Bartsch <bo@cedev.de>