[BUGFIX] Wrap action callback in Promise
The changes of #99176 and #99189 logically conflicted with each other.
The former resets the state of the DeferredAction by using `finally()`,
while the latter one introduces special handling when returning anything
that's not a Promise, leading to issues due to the bound `finally()`
call.

The callback execution itself is now resolved with a new Promise object,
making `finally()` functional again.

Resolves: #99195
Related: #99176
Related: #99189
Releases: main
Change-Id: I5507ad75b442b074fc477457ad23df49d6672d51
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/76824
Tested-by: core-ci <typo3@b13.com>
Tested-by: Benni Mack <benni@typo3.org>
Reviewed-by: Benni Mack <benni@typo3.org>