[TASK] Improve usability of "Create Administrative User" function
When the "Create Administrative User" is used in install tool, all
input fields are cleared regardless of, if a new user has been created
or not. This makes it impossible to correct possible validation errors
and the users has to re-enter all fields from scratch.

With this patch, input fields get only cleared, when the new user has
successfully been created.

Resolves: #99208
Releases: main
Change-Id: Ic04fc3daf394c2149ade2245b3c86b4c0b1f4067
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/76842
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Markus Klein <markus.klein@typo3.org>
Tested-by: Andreas Fernandez <a.fernandez@scripting-base.de>
Tested-by: core-ci <typo3@b13.com>
Reviewed-by: Andreas Fernandez <a.fernandez@scripting-base.de>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>