[TASK] Use unified buttons in Scheduler docheader
The 'SchedulerTask' module should use
same buttons in docheader as in other modules, therefore
use only 'Close', 'Save and Close', 'Save', 'New' and 'Delete'
and in the correct context.

Leave 'Save and Close' button for now,
until there exist an state handling for closing the form with changes.

Resolves: #99230
Releases: main
Change-Id: I907f7b93d3fa09b98bc830fe232c12557438db41
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/76867
Tested-by: Annett Jähnichen <mcmietz@web.de>
Tested-by: core-ci <typo3@b13.com>
Reviewed-by: Annett Jähnichen <mcmietz@web.de>
Reviewed-by: Stefan Bürk <stefan@buerk.tech>
Reviewed-by: Benni Mack <benni@typo3.org>
Tested-by: Stefan Bürk <stefan@buerk.tech>
Tested-by: Benni Mack <benni@typo3.org>