[FEATURE] Add minimum age option to cleanup:deletedrecords cli command
Using the cli command `cleanup:deletedrecords` to clean up the database
periodically is not really possible with an installed EXT:recycler,
because all records marked for deletion are deleted immediately and thus
the recycler seems less useful.

The new option `--min-age` added to the `cleanup:deletedrecords` cli
command allows to define a minimum age of X days a record needs to be
marked as deleted before it gets deleted.

Resolves: #99258
Releases: main
Change-Id: If9daa0e7266868ee14ad9282e4997031b85c80b6
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/76912
Reviewed-by: Anja Leichsenring <aleichsenring@ab-softlab.de>
Tested-by: Stefan Bürk <stefan@buerk.tech>
Tested-by: Anja Leichsenring <aleichsenring@ab-softlab.de>
Tested-by: core-ci <typo3@b13.com>
Reviewed-by: Stefan Bürk <stefan@buerk.tech>