[TASK] Align error responses in ReactionResolver
Error responses now use the same format in
ReactionResolver as well as in the reaction.

Additionally, a new method allows to cleanup
some duplicated code and the default status
code is 400, since 503 indicates a server error,
which is not the case for the current errors. All
of them are client errors: wrong identifier, invalid
payload, etc...

Resolves: #99309
Related: #98373
Releases: main
Change-Id: Iefa485c3a15778a69d99f365f9e0e7e8a3eb7e2f
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/76974
Tested-by: Benni Mack <benni@typo3.org>
Reviewed-by: Oliver Bartsch <bo@cedev.de>
Tested-by: Andreas Fernandez <a.fernandez@scripting-base.de>
Reviewed-by: Benni Mack <benni@typo3.org>
Tested-by: Oliver Bartsch <bo@cedev.de>
Reviewed-by: Andreas Fernandez <a.fernandez@scripting-base.de>
Tested-by: core-ci <typo3@b13.com>