[BUGFIX] Dynamically set 'clear' TypoScript flag
Not too many people know and understand the details
of the 'clear constants' and 'clear setup' flags in
TypoScript sys_template records: It not only
resets any given tree coming from previous templates,
but also triggers inclusion of TypoScript from
globals. As such, it is important to set the flags
somewhere, usually for the first template row. This
is also the main difference between the "Create
template for a new site" and "Create an extension
template" buttons: "New site" sets the clear flags
to trigger globals inclusion.

The old TypoScript parser had a convenience solution
to still include globals in case no template record
did set the 'clear' flags properly.

The new parser did not do that. We got feedback from
multiple early v12 adopters on this, who stumbled
here: If the clear flags are not set, especially the
basic rendering definitions from fluid_styled_content
are missing. Since this issue can be pretty hard to
find, the patch changes the SysTemplateTreeBuilder to
set the clear flag on the first template row if no
other row sets it. As a result, this increases b/w
compat with the old parser and integrators can
essentially forget about this flag for casual simple
setups again.

Resolves: #99331
Related: #97816
Related: #98552
Change-Id: I839e0fd04ff432911ff59bc6ec2047eb1e53be3e
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/77010
Tested-by: core-ci <typo3@b13.com>
Tested-by: André Kraus <info@andrekraus.com>
Reviewed-by: Anja Leichsenring <aleichsenring@ab-softlab.de>
Tested-by: Stefan Bürk <stefan@buerk.tech>
Reviewed-by: André Kraus <info@andrekraus.com>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Anja Leichsenring <aleichsenring@ab-softlab.de>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Stefan Bürk <stefan@buerk.tech>