[TASK] Improve TypoScript submodule naming
By renaming the submodules of the ext:tstemplate "TypoScript" backend
module and adding descriptive texts, it makes it clearer and more
beginner-friendly for what each submodule is used for.

The headings of the submodules now also correspond to the names of
corresponding option in the submodule selector.

By renaming words like "template" to "TypoScript template" a
unified wording will be used.

Furthermore, the submodule "TypoScript Records" is now set as default.

Resolves: #99330
Releases: main
Change-Id: Icc3aa61a4755cb96b03fdd134a7c7e47788107b8
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/77017
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Stefan Bürk <stefan@buerk.tech>
Tested-by: core-ci <typo3@b13.com>
Tested-by: Stefan Bürk <stefan@buerk.tech>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>