[BUGFIX] Use correct flag of Zimbabwe
The icon set used for the flags in TYPO3 accidentally used the flag of
Andorra for Zimbabwe. Unfortunately, it's not tracable anymore where
those flags came from, therefor, the flag of Zimbabwe is copied from
https://github.com/lipis/flag-icons and has been converted to PNG.

The flags of https://github.com/lipis/flag-icons may be used in the long
run as a maintained replacement.

Resolves: #99344
Releases: main, 11.5
Change-Id: I68e0e3b99254f35cce8c84a668a04765ea7b8159
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/77030
Tested-by: Torben Hansen <derhansen@gmail.com>
Tested-by: core-ci <typo3@b13.com>
Reviewed-by: Torben Hansen <derhansen@gmail.com>
Tested-by: Oliver Bartsch <bo@cedev.de>
Tested-by: Andreas Fernandez <a.fernandez@scripting-base.de>
Reviewed-by: Andreas Fernandez <a.fernandez@scripting-base.de>
Reviewed-by: Oliver Bartsch <bo@cedev.de>