[BUGFIX] Use correct element browser config in "Setup" module
The configuration for the element browser used for the avatar field in
the "Setup" module is now properly composed to be parseable by the
concrete `FileBrowser` implementation. The correct default upload folder
is now determinable.

Additionally, the `FileBrowser` now loads the folder's icon only if the
`$selectedFolder` property is correctly set to avoid type errors.

Resolves: #99367
Related: #97933
Releases: main
Change-Id: I15f9b67d5066f213ff7a7cb69f6f5f0f862d8e2b
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/77129
Tested-by: core-ci <typo3@b13.com>
Tested-by: Georg Ringer <georg.ringer@gmail.com>
Reviewed-by: Georg Ringer <georg.ringer@gmail.com>
Reviewed-by: Oliver Bartsch <bo@cedev.de>
Tested-by: Oliver Bartsch <bo@cedev.de>