[BUGFIX] Catch any transfer exception in HTTP host header check
ServerResponseCheck triggers a HTTP host header check which is
expected to fail. The more generic TransferException is used to
catch any other failed request, not only those with 4xx or 5xx
HTTP status codes. Besides that, TLS certificates shall not be
verified, and HTTP location redirects not be followed.

Resolves: #99368
Releases: main, 12.1, 11.5, 10.4
Change-Id: Id40457d4408c74d9229d4e6dcdedc0b69ffe9667
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/77153
Reviewed-by: Oliver Hader <oliver.hader@typo3.org>
Tested-by: Oliver Hader <oliver.hader@typo3.org>
Tested-by: core-ci <typo3@b13.com>