[BUGFIX] Fix absolute backend URL generation in frontend context
$normalizedParams->getRequestDir() returns '/' for frontend
requests and `/typo3/` for backend requests. This results in
problems when the backend UriBuilder relies on getRequestDir()
to provide the `/typo3/` suffix for backend URL generation.

UriBuilder is now changed to base absolute URLs on
getSiteUrl(), which is defined to return equal values
for backend and frontend requests, the typo3 suffix
is added manually.

Note that v12 introduced similar behavior with #99234,
where UriBuilder was adapted to use BackendEntryPointResolver,
which rebases the backend URL calculation on the site path as well.

Also note that the ABSOLUTE_URL mode in backend UriBuilder isn't
actually used by the TYPO3 frontend, but some extensions started
to execute system reports in frontend context, which
exposed this bug with the introduction of #99347.

Releases: 11.5, 10.4
Resolves: #99368
Related: #99347
Related: #99234
Change-Id: Ifaaeb4725c0243d34603dc86b2c89d12d9c06bdd
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/77160
Reviewed-by: Benjamin Franzke <bfr@qbus.de>
Reviewed-by: Oliver Hader <oliver.hader@typo3.org>
Tested-by: Benjamin Franzke <bfr@qbus.de>
Tested-by: core-ci <typo3@b13.com>
Tested-by: André Buchmann <andy.schliesser@gmail.com>
Tested-by: Oliver Hader <oliver.hader@typo3.org>
Reviewed-by: André Buchmann <andy.schliesser@gmail.com>