[BUGFIX] Always allow target attribute
TYPO3 included xhtml_11 (output doctype)
standard when the second revision of
XHTML 1.1 was not published yet.

TYPO3's code base assumed that in some
cases the "target" attribute (in XHTML 1.0 strict and XHTML 1.1)
was not allowed, which is not correct.

See http://learningtheworld.eu/2007/xhtml-with-target/

Additional reading: https://en.wikipedia.org/wiki/XHTML#XHTML_1.1:_Module-based_XHTML

For this reason, the workarounds for XHTML 1.0 strict
and XHTML 1.1 are removed.

Resolves: #99411
Releases: main
Change-Id: Iaf65aea10e191d24c0d78ee558028a6c63539754
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/77200
Tested-by: Georg Ringer <georg.ringer@gmail.com>
Reviewed-by: Oliver Bartsch <bo@cedev.de>
Tested-by: Oliver Bartsch <bo@cedev.de>
Reviewed-by: Oliver Klee <typo3-coding@oliverklee.de>
Tested-by: core-ci <typo3@b13.com>
Reviewed-by: Georg Ringer <georg.ringer@gmail.com>