[BUGFIX] Respect HTTP_REFERER for felogin redirect mode 'referer'
The `redirectMode=referer` configuration allows to process a redirect
after a successful login to an evaluated referer value.

For the legacy version of the felogin plugin, the referer value has
been evaluated using the GET/POST parameter `referer` and the
`HTTP_REFERER` as fallback.

The extbase version of the felogin plugin currently only evaluates the
GET/POST parameter `referer`.

This patch adds the missing `HTTP_REFERER` fallback evaluation
of the referer.

Resolves: #91844
Releases: main, 11.5
Signed-off-by: Torben Hansen <derhansen@gmail.com>
Change-Id: Id4119b0425ddca09a350f4d8d8a6ebb4d3b3135b
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/77204
Reviewed-by: Oliver Hader <oliver.hader@typo3.org>
Reviewed-by: Benni Mack <benni@typo3.org>
Tested-by: core-ci <typo3@b13.com>
Reviewed-by: Markus Klein <markus.klein@typo3.org>
Reviewed-by: Guido Schmechel <guido.schmechel@brandung.de>
Tested-by: Oliver Hader <oliver.hader@typo3.org>