[BUGFIX] Properly localize placeholder
Due to the introduction of the "language" TCA type,
the column processing of TCA language fields in
TcaSelectItems->addData was skipped and therefore
the array conversion through the "processDatabaseFieldValue"
method was no longer applied.

The placeholder processing in TcaInputPlaceholders
however expected the tables  language field to be
an array. This is now fixed.

Resolves: #99446
Releases: main, 11.5
Change-Id: I20f456ba7066c19d2eded00079e58c4c64521eeb
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/77265
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: core-ci <typo3@b13.com>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>