[TASK] Allow npm version 9.x for node builds
npm version 8.x and 9.x are officially supported for
nodejs 16. Normal npm updates updates from 8 to 9.
Currently npm 9 is disallowed by the version constraint
in the packages.json.

This change allows npm 9 along with npm 8 to be used
to work with npm and dependencies. This allows us to
rebuild the nodejs images without pinning npm down to
npm 8.x.

Resolves: #99480
Releases: main
Change-Id: I962b9da0fee6d99278e86e4e86ca2be604ed679c
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/77278
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Stefan Bürk <stefan@buerk.tech>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Stefan Bürk <stefan@buerk.tech>
Tested-by: core-ci <typo3@b13.com>