[FEATURE] Show the redirect integrity status
The `ext:redirects` provides a integrity check
command to detect integrity issues of existing
redirects.

This change persists the integrity state to the
redirect records and extends the backend module
to make the integrity state of redirects visible
by colorizing them in the list. Additionally, a
new filter option for the integrity state is now
available.

Resolves: #99485
Releases: main
Change-Id: I03db4472f3e37594e33ab92a617a3787014e157a
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/77285
Reviewed-by: Anja Leichsenring <aleichsenring@ab-softlab.de>
Tested-by: Stefan Bürk <stefan@buerk.tech>
Tested-by: core-ci <typo3@b13.com>
Reviewed-by: Stefan Bürk <stefan@buerk.tech>
Tested-by: Anja Leichsenring <aleichsenring@ab-softlab.de>