[BUGFIX] Cover all CKEditor4 toolbar configurations in migration
CKEditor4 had several ways of defining toolbars - Basic, Advanced,
Toolbar Groups, and Toolbar References. All these cases are now covered
in the migration and updated to the CKEditor5 format.

There can only be one `toolbar` or a `toolbarGroups` definition. The
preference is `toolbarGroups` since we assume it only exists on an
incomplete migration.

Instead of mapping groups to new definitions, we are now resolving each
main and subgroup to the CKEditor4 defaults. Each item is migrated to
the CKEditor5 equivalent if one exists.

CKEditor5 allows buttons to appear multiple times in a configuration.
The migration now ensures that buttons exist once. Duplicates are
removed during the migration to reproduce the behavior of CKEditor4.

Resolves: #99556
Releases: main

Change-Id: Ib403f377895cc0f8020b7f0070a969088a9eec28
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/77358
Reviewed-by: Andreas Fernandez <a.fernandez@scripting-base.de>
Tested-by: Benni Mack <benni@typo3.org>
Tested-by: core-ci <typo3@b13.com>
Reviewed-by: Benni Mack <benni@typo3.org>