[TASK] Streamline flag icons
We streamlined the flag icons and make them easier to handle.
The core provides a range of flag icons that are representing countries,
regions, movements, islands, and more. The flags are mostly used in
conjunction with languages.

We agree that a flag does not represent a language, but having a visual
identifier attached to languages makes it easier for editors to
identify the language they want to edit or translate.

New flags added in this patch will express that we understand both,
the issue and the need to differentiate languages. We chose simple
colored flags to achieve this. It still allows differentiation while
variants like de-DE and de-CH can be maintained and identified.

New flags: black, blue, cyan, green, indigo, orange, pink, purple, red,
teal, white, yellow, rainbow.

Flags of historic countries have been removed:
- AN, Netherlands Antilles (until 2010)
- CS, State Union of Serbia and Montenegro (until 2006)

Flags for language codes have been removed:
- kl, Greenlandic
- mi, Māori

Flags for country regions have been aligned:
- Spain, Catalonia: catalonia -> es-ct
- Canada, Quebec: qc -> ca-qc

Please adjust your site configuration if you are using one of
the removed or renamed flag icons.

Releases: main
Related: #90564
Related: #87062
Resolves: #99609
Change-Id: Id7d3c3358c514afda3f7c0553eaca59c2220f226
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/77472
Reviewed-by: Stefan Bürk <stefan@buerk.tech>
Tested-by: Stefan Bürk <stefan@buerk.tech>
Tested-by: core-ci <typo3@b13.com>
Tested-by: Andreas Fernandez <a.fernandez@scripting-base.de>
Reviewed-by: Andreas Fernandez <a.fernandez@scripting-base.de>