[TASK] Use official translations for ISO 639-1/2 languages
Previously, TYPO3 manually had lists of
labels for languages (primarily used to
select a user's language for the TYPO3 Backend
interface).

Similar to the Country API, the list
of official languages are now fetched
from a debian repository, which provides
a list of official language keys and
their labels so translators do not need
to update this static list, and developers
do not need to add a label when a new language
is added.

A list of two-letter code language keys
in addition to three-letter code language keys
are now available and can slowly be used
to be supported instead of the TYPO3 core list.

Previous "pseudo-classes" such as IsoCodeService
and their respective XLF files are removed.

Code in be_users.lang itemsProcFunc and
SetupModuleController can then be simplified.

Resolves: #100659
Releases: main
Change-Id: I26dcf24d496856e5a2a2c9c0b27193be724d7df1
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/77503
Reviewed-by: Markus Klein <markus.klein@typo3.org>
Reviewed-by: Benni Mack <benni@typo3.org>
Tested-by: Georg Ringer <georg.ringer@gmail.com>
Reviewed-by: Georg Ringer <georg.ringer@gmail.com>
Tested-by: core-ci <typo3@b13.com>
Tested-by: Markus Klein <markus.klein@typo3.org>
Tested-by: Benni Mack <benni@typo3.org>