[TASK] Define explicit routes for Extbase Backend Modules
Previously, the controller / action pairs of an Extbase Backend
Module were defined via corresponding GET parameters. Thanks to
the new Module Registration API and the support for individual
sub routes, it's now also possible to define explicit routes
for each controller / action pair. This is done automatically,
as long as the "enableNamespacedArgumentsForBackend" feature
toggle is turned off, which is the default.

This therefore results in following change:

http://example.com/typo3/module/system/BeuserTxBeuser?controller=BackendUser&action=filemounts

becomes

http://example.com/typo3/module/system/BeuserTxBeuser/BackendUser/filemounts

Resolves: #99704
Related: #99647
Related: #96733
Releases: main
Change-Id: Ie7bbf70793f7e3da17db3ab1a322ba8ad7bcc5b8
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/77508
Reviewed-by: Oliver Bartsch <bo@cedev.de>
Reviewed-by: Benni Mack <benni@typo3.org>
Tested-by: Stefan Bürk <stefan@buerk.tech>
Reviewed-by: Stefan Bürk <stefan@buerk.tech>
Tested-by: core-ci <typo3@b13.com>
Tested-by: Oliver Bartsch <bo@cedev.de>
Tested-by: Benni Mack <benni@typo3.org>