[BUGFIX] Do not access tce_db route directly
In case only one cache clear option is available
in the toolbar (no dropdown), clicking on the icon
previously led to a blank page, since the corresponding
JS event listener did not get called and the tce_db
route was called directly.

This is now fixed by using the correct class to notify
the event listener, which will call tce_db in an ajax request.

Resolves: #99671
Releases: main, 11.5
Change-Id: I0f6e9259fdf189b5fd021cdc03d9e5d57a290e01
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/77533
Tested-by: Oliver Bartsch <bo@cedev.de>
Reviewed-by: Oliver Bartsch <bo@cedev.de>
Tested-by: Stefan Bürk <stefan@buerk.tech>
Tested-by: core-ci <typo3@b13.com>
Reviewed-by: Stefan Bürk <stefan@buerk.tech>
Reviewed-by: Benni Mack <benni@typo3.org>
Tested-by: Benni Mack <benni@typo3.org>