[BUGFIX] Restore backend module rerouting functionality
Due to the changes in #99620, it is now possible
for a second level module to not define any route.
Such module then automatically uses the routes
from the first available sub module (third-level).

However, this now led to the fact that it was no
longer possible for the BackendModuleValidator
to restore the actually last used third-level module
of such sub modules, because the module object
in the Route object got replaced by the fallback.

This use case is now also properly handled by
checking whether the resolved module is the
request modules' fallback.

Resolves: #99684
Related: #99620
Releases: main
Change-Id: I8cd19417fa66593dd4c612a07723dcbcf5162f5f
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/77539
Reviewed-by: Stefan Bürk <stefan@buerk.tech>
Tested-by: core-ci <typo3@b13.com>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Stefan Bürk <stefan@buerk.tech>