[TASK] Adjusting the appearance of the tables in the list module
To provide a streamlined look&feel of record tables in the record list,
their table header's background color now matches those of panels, e.g.
the "Clipboard" at the very bottom. For the same reason, a box-shadow is
applied.

To achieve a better visual separation, the table margin is increased.

Resolves: #99727
Releases: main
Change-Id: I7a64a6c2e3373f956a506630741628b20729b7a4
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/77611
Tested-by: Jasmina Ließmann <minapokhalo+typo3@gmail.com>
Tested-by: core-ci <typo3@b13.com>
Reviewed-by: Jasmina Ließmann <minapokhalo+typo3@gmail.com>
Reviewed-by: Benni Mack <benni@typo3.org>
Tested-by: Benni Mack <benni@typo3.org>