[TASK] Drop internal annotation from Extbase's UriBuilder->setRequest()
The setRequest() method is necessary to be called to set the current
request object when using the UriBuilder outside of an Extbase
controller or a view helper. The according changelog also states in
the migration section that the setRequest() method should be called.
So, we remove the internal annotation, so the method is now considered
as public API.

Related: #99650
Resolves: #99734
Releases: main
Change-Id: I45a2149d3215a7ea9abb2458da5eba922bed4cc7
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/77630
Reviewed-by: Benni Mack <benni@typo3.org>
Tested-by: Torben Hansen <derhansen@gmail.com>
Reviewed-by: Torben Hansen <derhansen@gmail.com>
Tested-by: core-ci <typo3@b13.com>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>