[TASK] Further refactorings of SVG tree drag+drop code
This change is the next iteration
of cleaning up SVG-drag/drop code.

Handling of the "positioning line" to drag nodes
between other nodes (page tree only)
has been encapsulated in drag-drop.ts

A lot of methods and variables are renamed and comments
are added in order to understand the code better.

A few methods regarding draggables (the shadowed content) have
been centralized so the code is easier to
read where CSS classes are modified.

Resolves: #99783
Releases: main
Change-Id: I9e28d0ca3d886afaa8b3e2cf274593d3e45c3fa5
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/77656
Tested-by: Andreas Fernandez <a.fernandez@scripting-base.de>
Tested-by: core-ci <typo3@b13.com>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Andreas Fernandez <a.fernandez@scripting-base.de>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>