[TASK] Performance improvement in 'Active TypoScript'
Since #99782, we're always rendering the entire
TypoScript tree. With bigger trees like a loaded
bootstrap_package, these can be thousands of nodes.

One bummer is the f:be.link VH call for each node.
The patch suppresses this with a pretty straight
hack, improving server side by about 40%.

Another optimization (accessing object members)
requires a change in Typo3Fluid, which needs to
be prepared and released on this side first.

Resolves: #99820
Related: #99782
Releases: main
Change-Id: I688d1038a5e4f23c5f2172d64d7508cdda3e3ff2
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/77695
Reviewed-by: Jasmina Ließmann <minapokhalo+typo3@gmail.com>
Reviewed-by: Anja Leichsenring <aleichsenring@ab-softlab.de>
Tested-by: Anja Leichsenring <aleichsenring@ab-softlab.de>
Tested-by: core-ci <typo3@b13.com>
Tested-by: Jasmina Ließmann <minapokhalo+typo3@gmail.com>